Add option to provide phantom with a random port to talk with phantomjs
phantom talks with phantomjs by starting a http server.
In order to ensure that all testreduce clients don't start
this http server on the same port (which seems to be happening
oddly enough despite phantom asking the OS for a free port),
assign a random port.

Without this option, when I start N testreduce clients, most of
them seem to start a http server on the same exact port and has
phantomjs processes listening on that port. All except one of these
phantomjs actually survives the port conflict and the rest of the
testreduce clients hang .. so in effect, most of the testreduce
clients are worthless.

See the link below for how phantom talks with phantomjs
https://github.com/amir20/phantomjs-node/tree/v1#no-really-how-does-it-work

Change-Id: Id08d7d32ba67ef0ee635ec1efd976a945fa1ba3f