Add retries to uprightdiff with a different block size
uprightdiff is crashing (segv) on some tests with fairly minimal
diffs. The crasher disappears when I reverse the order of images.
It also disappears by changing the block size from the default (16).

For now, add a single retry with a block size of 8 which lets many
of this previously crashing tests to run to completion.

Change-Id: Iccfb374a2dbd4d3eb136757cb3024318992e1135