Strip p-wrappers around "empty" paragraphs
These are all introduced discussion-tool comments from extensions
which then first triggers a span wrapper around the comment to
preserve about-id continuity, which in turn triggers a p-wrapper
because empty-pargraph detection fails to kick in.

In turn, all these spurious comments are added after every single
templatestyle (and possibly other extension uses) present on a
talk page (directly or via a template) because Parsoid processes
them as top-level docs and runs the 'onParserAfterTidy' hook which
DiscussionTools registers a handler for.

Bit of a mess, but, we may have to get DiscussionTools to adapt
itself for Parsoid land a bit in this case.

But, for now, let us suppress spurious vertical whitespace during
visual diff tests that gets in the way of identifying other
breakages.

Change-Id: I6496a47be1594295af50fc83def754ba44b62cc3