Add canvas-based diff image browser
It's a little hard to test, but it's a one-page standalone app.  If
you load the browser.html file as a file:/// url locally, and then
just edit the <img src> in that file to point to actual images (even
https:// urls pointing to the real visual diff server) you should be
able to test it.

It's not particularly "beautiful".  But it's functional, and it lets
you scroll-wheel up and down visual diffs, drag and click to pan them
around, click to cycle between the various views (right click to cycle
the opposite direction), and shift-drag will let you move the parsoid
layer independent of the core layer if you want to realign a section
at the bottom of the page.

UX can be improved, but it's a starting point.

Bug: T383166
Change-Id: Ic15f1b84a6737cc1f37bc390e8ee6f81dbb92ce0