Update mobile handling for more reliable screenshots
This makes some changes which should better represent mobile editors
and confidence in the mobile version


1) Make sure we emulate to load correct CSS and JS assets. Using
useformat and useskin is not the same as using a mobile device as
it doesn't load the same site JS/CSS.
2) Use mobile resolution and expand the first two sections. Since most mobile users are viewing the site this is more realistic of the experience we want to protect. Since we have visual diff tests for
desktop that should be sufficient for the expanded case.

Additional changes:
Mobile logic is moved behind a `mobile` flag to centralize it
for future.


Follow up to 423d28c2843

Change-Id: I237b8a261408cc4c9a6e95cfa2dc19228b9322c1